// src/app/Model/Experience.ts
import { EmploymentType } from './EmploymentType';

export class Experience
{
    id:number;
    summary: string;
    title: string;
    EmploymentType: EmploymentType;
    Name: string;
    location: string;
    startDate: Date;
    endDate: Date;  
     headLine: string;
     Description: string;
}